requirements.txt
bson
certifi
flask
flask-jwt-extended
flask-restful
flask-restx
flasgger
numpy
openai
pandas
pymongo
python-dotenv
Pillow